tests: classWithoutStereotype
